refactor codebase structure
