+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
-package-lock.json
 new.js
 errs.js
 tokens.js
 test.js
+.env
