+++ example.env
@@ -1,0 +1,7 @@
+PORT=3000
+PAGE_ACCESS_TOKEN='your page access token'
+FACEBOOK_GRAPH_API_URL='https://graph.facebook.com/v2.6/'
+SB_URL='https://your.url'
+SB_KEY='your-key'
+RENDER_EXTERNAL_URL='https://your.url'
+MYSERVER='server'
