+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node src/index"
   },
   "keywords": [],
   "author": "",
@@ -13,6 +13,7 @@
     "@supabase/supabase-js": "^2.39.7",
     "axios": "^1.6.7",
     "botly": "^1.6.2",
+    "dotenv": "^16.4.5",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
     "socks-proxy-agent": "^8.0.3"
