+++ src/routes/index.js
@@ -1,0 +1,43 @@
+const os = require("os");
+const router = require('express').Router();
+
+const { formatBytes } = require('../utils');
+
+router.get("/", (req, res) => {
+  const memoryUsage = process.memoryUsage();
+  let uptimeInSeconds = process.uptime();
+  
+  let uptimeString = "";
+  if (uptimeInSeconds < 60) {
+    uptimeString = `${uptimeInSeconds.toFixed()} seconds`;
+  } else if (uptimeInSeconds < 3600) {
+    uptimeString = `${(uptimeInSeconds / 60).toFixed()} minutes`;
+    } else if (uptimeInSeconds < 86400) {
+      uptimeString = `${(uptimeInSeconds / 3600).toFixed()} hours`;
+    } else {
+      uptimeString = `${(uptimeInSeconds / 86400).toFixed()} days`;
+    }
+  
+    const osInfo = {
+      totalMemoryMB: (os.totalmem() / (1024 * 1024)).toFixed(2),
+      freeMemoryMB: (os.freemem() / (1024 * 1024)).toFixed(2),
+      cpus: os.cpus(),
+    };
+  
+    res.render("index", { memoryUsage, uptimeString, formatBytes, osInfo });
+});
+
+// TODO: add the events
+
+router.post("/webhook", (req, res) => {
+  if (req.body.message) {
+    onMessage(req.body.message.sender.id, req.body.message);
+  } else if (req.body.postback) {
+    onPostBack(req.body.postback.message.sender.id, req.body.postback.message, req.body.postback.postback);
+  }
+  res.sendStatus(200);
+});
+
+router.get("/ping", (req, res) => { res.status(200).json({ message: "Ping successful" }); });
+
+module.exports = router;
